+++ example/lib/main.dart
@@ -51,6 +51,7 @@ class _ExampleAppBodyState extends State<ExampleAppBody> {
         signatureType: useEc ? IOSSignatureType.ECDSA : IOSSignatureType.RSA,
         biometryCurrentSet: true,
       ),
+      enforceBiometric: true,
     );
     setState(() => keyMaterial = result);
     if (result != null) {
@@ -100,7 +101,8 @@ class _ExampleAppBodyState extends State<ExampleAppBody> {
         payload: payload!,
         promptMessage: 'Sign Payload',
         androidOptions: const AndroidSignatureOptions(
-          allowDeviceCredentials: true,
+          allowDeviceCredentials: false,
+          subtitle: 'Approve the login to continue',
         ),
         iosOptions: const IosSignatureOptions(shouldMigrate: false),
       ),